requirement.txt
langchain==0.1.5
langchain-community==0.0.19
langchain-core==0.1.19
streamlit==1.29.0
sentence-transformers==2.2.2
pypdf==3.17.4
chromadb==0.4.22
pydantic==2.5.3
requests==2.31.0
streamlit==1.32.0
numpy==1.26.0
langchain-core==0.1.18
langchain-community==0.0.16
chromadb==0.4.22
pytesseract==0.3.10
Pillow==10.1.0
PyMuPDF==1.23.8
regex==2023.10.3
langchain==0.1.0